+++ lib/src/actions.dart
@@ -3,60 +3,87 @@ import 'package:rltut/src/engine.dart';
 import 'package:rltut/src/entity.dart';
 
 abstract class Action {
-  void perform(Engine engine, Entity entity);
+  final Actor _entity;
+  Actor get entity => _entity;
+  Engine get engine => _entity.gameMap.engine;
+
+  Action(this._entity);
+
+  void perform();
+}
+
+class WaitAction extends Action {
+  @override
+  void perform() {}
+
+  WaitAction(Actor entity) : super(entity);
 }
 
 abstract class ActionWithDirection extends Action {
+  Vec _direction;
   final int _dx;
   final int _dy;
 
+  Vec get direction => _direction;
   int get dx => _dx;
   int get dy => _dy;
 
-  ActionWithDirection(this._dx, this._dy);
+  Vec get destination => entity.pos + direction;
+  Entity get blockingEntity =>
+      engine.gameMap.getBlockingEntityAtLocation(destination);
+  Actor get targetActor => engine.gameMap.getActorAtLocation(destination);
+
+  ActionWithDirection(Actor entity, this._dx, this._dy) : super(entity) {
+    _direction = Vec(_dx, _dy);
+  }
 
   @override
-  void perform(Engine engine, Entity entity);
+  void perform();
+}
+
+class MeleeAction extends ActionWithDirection {
+  MeleeAction(Actor entity, int dx, int dy) : super(entity, dx, dy);
+
+  @override
+  void perform() {
+    var target = targetActor;
+    if (target == null) return;
+
+    var damage = entity.fighter.power - target.fighter.defense;
+
+    var attackDesc = '${entity.name} attacks ${target.name}';
+
+    if (damage > 0) {
+      print('$attackDesc for $damage hit points.');
+      target.fighter.hp -= damage;
+    } else {
+      print('$attackDesc but does no damage.');
+    }
+  }
 }
 
 class MovementAction extends ActionWithDirection {
-  MovementAction(int dx, int dy) : super(dx, dy);
+  MovementAction(Actor entity, int dx, int dy) : super(entity, dx, dy);
 
   @override
-  void perform(Engine engine, Entity entity) {
-    var dest = entity.pos + Vec(dx, dy);
-
-    if (!engine.gameMap.inBounds(dest)) return;
-    if (!engine.gameMap.tiles[dest].walkable) return;
-    if (engine.gameMap.getBlockingEntityAtLocation(dest) != null) return;
+  void perform() {
+    if (!engine.gameMap.inBounds(destination)) return;
+    if (!engine.gameMap.tiles[destination].walkable) return;
+    if (engine.gameMap.getBlockingEntityAtLocation(destination) != null) return;
 
     entity.move(dx, dy);
   }
 }
 
-class MeleeAction extends ActionWithDirection {
-  MeleeAction(int dx, int dy) : super(dx, dy);
-
-  @override
-  void perform(Engine engine, Entity entity) {
-    var dest = entity.pos + Vec(dx, dy);
-    var target = engine.gameMap.getBlockingEntityAtLocation(dest);
-    if (target == null) return;
-
-    print('You kick the ${target.name}!');
-  }
-}
-
 class BumpAction extends ActionWithDirection {
-  BumpAction(int dx, int dy) : super(dx, dy);
+  BumpAction(Actor entity, int dx, int dy) : super(entity, dx, dy);
 
   @override
-  void perform(Engine engine, Entity entity) {
-    var dest = entity.pos + Vec(dx, dy);
-    if (engine.gameMap.getBlockingEntityAtLocation(dest) != null) {
-      return MeleeAction(dx, dy).perform(engine, entity);
+  void perform() {
+    if (targetActor != null) {
+      return MeleeAction(entity, dx, dy).perform();
     } else {
-      return MovementAction(dx, dy).perform(engine, entity);
+      return MovementAction(entity, dx, dy).perform();
     }
   }
 }