+++ lib/src/engine.dart
@@ -1,23 +1,40 @@
-import 'package:malison/malison.dart';
-import 'package:malison/malison_web.dart';
+import 'package:piecemeal/piecemeal.dart';
 
-import 'package:rltut/src/actions.dart';
 import 'package:rltut/src/entity.dart';
 import 'package:rltut/src/fov.dart';
 import 'package:rltut/src/gamemap.dart';
+import 'package:rltut/src/screens.dart';
+
+final moveKeys = {
+  'n': Vec(0, -1),
+  's': Vec(0, 1),
+  'w': Vec(-1, 0),
+  'e': Vec(1, 0),
+  'ne': Vec(1, -1),
+  'se': Vec(1, 1),
+  'nw': Vec(-1, -1),
+  'sw': Vec(-1, 1),
+};
 
 class Engine {
-  final GameMap _gameMap;
-  final Entity _player;
-  final Fov _fov;
+  final Actor _player;
+  GameMap _gameMap;
+  Fov _fov;
+  GameScreen _screen;
 
+  Actor get player => _player;
   GameMap get gameMap => _gameMap;
-  Entity get player => _player;
   Fov get fov => _fov;
+  GameScreen get screen => _screen;
 
-  Engine(this._gameMap, this._player, this._fov) {
-    updateFov();
+  set gameMap(GameMap value) => _gameMap = value;
+  set fov(Fov value) => _fov = value;
+  set screen(GameScreen value) => _screen = value;
+
+  Engine(this._player) {
+    _screen = MainGameScreen(this);
   }
+
   void updateFov() {
     fov.refresh(player.pos);
     for (var pos in gameMap.bounds) {
@@ -28,56 +45,12 @@ class Engine {
   }
 
   void handleEnemyTurn() {
-    for (var i = 1; i < gameMap.entities.length; i++) {
-      print('The ${gameMap.entities[i].name} hangs around, doing nothing.');
+    var enemies = gameMap.actors;
+    enemies.remove(player);
+    for (var entity in enemies) {
+      if (entity.ai != null) {
+        entity.ai.perform();
+      }
     }
-  }
-}
-
-class GameScreen extends Screen<String> {
-  final Engine _engine;
-
-  Action _action;
-
-  Engine get engine => _engine;
-
-  GameScreen(this._engine);
-
-  @override
-  bool handleInput(String input) {
-    switch (input) {
-      case 'n':
-        _action = BumpAction(0, -1);
-        break;
-      case 's':
-        _action = BumpAction(0, 1);
-        break;
-      case 'w':
-        _action = BumpAction(-1, 0);
-        break;
-      case 'e':
-        _action = BumpAction(1, 0);
-        break;
-      default:
-        return false;
-    }
-
-    _action.perform(engine, engine.player);
-    engine.handleEnemyTurn();
-    engine.updateFov();
-
-    ui.refresh();
-
-    return true;
-  }
-
-  @override
-  void update() {
-    dirty();
-  }
-
-  @override
-  void render(Terminal terminal) {
-    engine.gameMap.render(terminal);
   }
 }