+++ lib/src/renderorder.dart
@@ -1,0 +1,1 @@
+enum RenderOrder { corpse, item, actor }